refactor: move exported helpers out of internal
To match the rule that exports should be in the public file structure we
move out those files from internal to a top level `./helpers` folder.
