+++ v-next/hardhat-ignition/test/deploy/rerun/rerun-a-deploy-that-timed-out.ts
@@ -5,7 +5,7 @@ import { assertHardhatInvariant } from "@ignored/hardhat-vnext-errors";
 import { buildModule, wipe } from "@ignored/hardhat-vnext-ignition-core";
 import { assert } from "chai";
 
-import { HardhatArtifactResolver } from "../../../src/hardhat-artifact-resolver.js";
+import { HardhatArtifactResolver } from "../../../src/helpers/hardhat-artifact-resolver.js";
 import { mineBlock } from "../../test-helpers/mine-block.js";
 import { useFileIgnitionProject } from "../../test-helpers/use-ignition-project.js";
