+++ v-next/hardhat-ignition/test/deploy/reset.ts
@@ -2,7 +2,7 @@ import { assertHardhatInvariant } from "@ignored/hardhat-vnext-errors";
 import { status } from "@ignored/hardhat-vnext-ignition-core";
 import { assert } from "chai";
 
-import { HardhatArtifactResolver } from "../../src/hardhat-artifact-resolver.js";
+import { HardhatArtifactResolver } from "../../src/helpers/hardhat-artifact-resolver.js";
 import { useFileIgnitionProject } from "../test-helpers/use-ignition-project.js";
 
 describe("reset flag", function () {
