+++ v-next/hardhat-ignition/test/utils/error-deployment-result-to-exception-message.ts
@@ -8,7 +8,7 @@ import type {
 import { DeploymentResultType } from "@ignored/hardhat-vnext-ignition-core";
 import { assert } from "chai";
 
-import { errorDeploymentResultToExceptionMessage } from "../../src/internal/utils/error-deployment-result-to-exception-message.js";
+import { errorDeploymentResultToExceptionMessage } from "../../src/helpers/error-deployment-result-to-exception-message.js";
 
 describe("display error deployment result", () => {
   describe("validation", () => {
