+++ v-next/hardhat-ignition/test/utils/resolve-deployment-id.ts
@@ -1,6 +1,6 @@
 import { assert } from "chai";
 
-import { resolveDeploymentId } from "../../src/internal/utils/resolve-deployment-id.js";
+import { resolveDeploymentId } from "../../src/helpers/resolve-deployment-id.js";
 
 describe("deploy id rules", () => {
   const exampleChainId = 31337;
